haciendo la parte de pagos

+++ src-tauri/src/mods/error.rs
@@ -7,7 +7,7 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum AppError {
     #[error("Error de monto, el monto a pagar es: {a_pagar:?},el monto pagado es: {pagado:?}")]
-    AmountError { a_pagar: f64, pagado: f64 },
+    AmountError { a_pagar: f32, pagado: f32 },
     #[error("Error de {0}")]
     IncorrectError(String),
     #[error("Solo existen dos posiciones para venta")]
